feat(content): Add endpoint to fetch content by owner id
A new API endpoint has been added that allows fetching of content based on the owner's id. This includes changes in the metadata.rs file where a new function `get_contents_by_owner` is defined. The route for this function is also set up in routes.rs. Additionally, the 'owner_id' field in the Program struct (in program.rs) has been changed from ObjectId to i32 to match with the newly created endpoint parameter type.